Amendment actioningTime promise magic
Renamed enum ServiceEndTime to ServiceEndState, as in API (since it is
module-wide and needs to ensure uniqueness)
Revised how promises are used, so that methods return promises through
a consistent route, whether it be a thin promise or an action that
genuinely needs deferring.
Made (not yet tested) magic methods to apply actioning time to
amendments.
Made stub for UpdateComponentValueAmendment.
Made a magic fetchIfNecessary() helper on BillingEntity.